Fix modal dialogs on small screens
On small laptop screens the relative positioning of the modal dialogs
could push the larger dialogs out of the screen without showing
scrollbars. This commits makes the body to be always 800 pixels tall. I
tried basing this over the tallest dialog I could locate but due to
unfamiliarity with the application can't be 100% I got it right.
However, when a larger size is needed it's easy to increase the pixel
count.

I tried testing the commit with capybara-webkit but the resize_window
functionality seems to have an open issue in their tracker about not
working properly.