Security update rails_admin
rails_admin says that versions until 0.4.3 are vulnerable to the YAML
parsing problem that was also a problem for Rails. This is only a
problem for serialized fields and hhj does not use them so we shouldn't
be currently vulnerable. I only realized that after upgrading but it's
still good practise to keep things updated.

I had troubles loading old sessions so the commit just changes the
secret token so that old sessions get discarded. To get the upgrade play
together with other gems they needed upgrades (like mongoid) which had
ripple effects.