Rails 4.0 upgrade
Upgrade rails to 4.0 and the gems needed to fix version conflicts.
Mongoid went to version 4.0 which meant that their JSON output changed.
id now is a hash with $oid as the key.

Factories specify the id. In previous versions this silently worked. Now
it started to throw an exception which meant that the code needed to be
adjusted to properly use the singletons. This was the easier approach
than changing all test logic to match different fixture generation.